fix(arc-676): move mediaInfo type to shared, use global media types
