+++ src/modules/shared/components/MediaCard/MediaCard.types.ts
@@ -1,5 +1,7 @@
 import { MouseEvent, ReactNode } from 'react';
 
+import { MediaTypes } from '@shared/types';
+
 export type MediaCardViewMode = 'list' | 'grid';
 
 export interface MediaCardProps {
@@ -11,7 +13,7 @@ export interface MediaCardProps {
 	publishedAt?: Date;
 	publishedBy?: string;
 	title?: string;
-	type?: 'video' | 'audio';
+	type: MediaTypes;
 	view?: MediaCardViewMode;
 	onBookmark?: (evt: MouseEvent) => void;
 	actions?: ReactNode;
